Correct GitHub username used in URLs
I guess the ownership of the repository was handed over from timedata-org to
rec, as GitHub serves a redirect from the former user to the latter.

This corrects the URLs used in the README and change log to use the new GitHub
username.

A big benefit of this, is the link to the documentation works again.